Contador especia terminado

+++ src_files/src/src_cliente/client_comandos/cmd_modificar_especia.h
@@ -4,9 +4,9 @@
 #include "client_comando.h"
 
 class CmdModificarEspeciaServer : public ComandoCliente {
-    uint8_t cantidad_especia;
+    uint16_t cantidad_especia;
 public:
-    CmdModificarEspeciaServer(uint8_t cantidad_especia);
+    CmdModificarEspeciaServer(uint16_t cantidad_especia);
 
     virtual bool ejecutar(WorldView& worldView) const override;
